Rename `result_ext` to `result`
